Cleanup nits
